strict_modular_configuration/pyproject.toml
[project]
name = "snippet"
version = "0.0.1"
dependencies = [
    "pydantic",
    "pyyaml"
]
requires-python = ">=3.8"
description = "Snippet package used to demonstrate something. Refer to README for more info."
readme = "README.md"


[project.optional-dependencies]
dev = ["black", "isort", "pytest"]

[project.entry-points."spam.magical"]
tomatoes = "spam:main_tomatoes"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["snippet/"]
